+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Vault/New-AzDataProtectionBackupVault.ps1
@@ -49,7 +49,7 @@ function New-AzDataProtectionBackupVault
         [ValidateSet('Disabled','Unlocked', 'Locked')]
         ${ImmutabilityState},
 
-        [Parameter(Mandatory=$false, HelpMessage='Cost Management Granularity of the vault. Allowed values are VaultLevel, ProtectedItemLevel, ProtectedItemWithParentTag.')]
+        [Parameter(Mandatory=$false, HelpMessage='Cost Management Granularity of the vault. Allowed values are VaultLevel, ProtectedItemLevel, ProtectedItemWithParentTag. Defaults to VaultLevel.')]
         [System.String]
         [ValidateSet('VaultLevel', 'ProtectedItemLevel', 'ProtectedItemWithParentTag')]
         ${CostManagementGranularity},
@@ -162,7 +162,10 @@ function New-AzDataProtectionBackupVault
         $hasCmkInfrastructureEncryption = $PSBoundParameters.Remove("CmkInfrastructureEncryption")
 
         $hasCostManagementGranularity = $PSBoundParameters.Remove("CostManagementGranularity")
-        if ($hasCostManagementGranularity) { $PSBoundParameters.Add("CostManagementSettingGranularityLevel", $CostManagementGranularity) }
+        if (-not $hasCostManagementGranularity) {
+            $CostManagementGranularity = "VaultLevel"
+        }
+        $PSBoundParameters.Add("CostManagementSettingGranularityLevel", $CostManagementGranularity)
 
         if (-not $hasCmkEncryptionState -and -not $hasCmkIdentityType -and -not $hasCmkUserAssignedIdentityId -and -not $hasCmkEncryptionKeyUri) {
             Az.DataProtection.Internal\New-AzDataProtectionBackupVault @PSBoundParameters